- now compile gcc with -Wstrict-prototypes as well as -Wall - fixes to eliminate more compiler warnings (at least for i386/FreeBSD JIT;   others probably remain when compiling under other configurations). - remove libraries/clib/native/ResourceBundle.[ch], they're not needed

+++ libraries/clib/awt/X/tlk.c
@@ -51,7 +51,7 @@ xErrorHandler ( Display *dsp, XErrorEvent *err )
   XGetErrorDatabaseText( dsp, "XRequest", key, "", buf, sizeof( buf));
   fprintf( stderr, "  request:    %s\n", buf);
 
-  fprintf( stderr, "  resource:   %X\n", err->resourceid);
+  fprintf( stderr, "  resource:   %X\n", (unsigned int) err->resourceid);
 
   return 0;
 }
